Classes da camada repository voltadas para interface
