+++ app/src/main/java/com/rafaelpereiraramos/testeandroidv2/view/statement/StatementsActivityViewModel.kt
@@ -4,17 +4,17 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.ViewModel
 import com.rafaelpereiraramos.testeandroidv2.db.model.StatementTO
-import com.rafaelpereiraramos.testeandroidv2.repo.ParameterRepo
+import com.rafaelpereiraramos.testeandroidv2.repo.impl.ParameterRepoImpl
 import com.rafaelpereiraramos.testeandroidv2.repo.ResourceWrapper
-import com.rafaelpereiraramos.testeandroidv2.repo.StatementRepo
+import com.rafaelpereiraramos.testeandroidv2.repo.impl.StatementRepoImpl
 import javax.inject.Inject
 
 /**
  * Created by Rafael P. Ramos on 27/11/2018.
  */
 class StatementsActivityViewModel @Inject constructor(
-    private val statementRepo: StatementRepo,
-    private val parameterRepo: ParameterRepo
+    private val statementRepo: StatementRepoImpl,
+    private val parameterRepo: ParameterRepoImpl
 ) : ViewModel() {
 
     private val _statements = MediatorLiveData<ResourceWrapper<List<StatementTO>>>()
